fix: make doc-fetch temp file and disk-space check build on MSVC
Two Windows/MSVC build breaks under /WX:

- otf_handler.cc used POSIX mkstemp/write/close/ssize_t and a hardcoded
  /tmp path for the doc-preview content fetch; <unistd.h> is non-Windows
  only, so these were undeclared on MSVC. Replace with a portable
  std::filesystem temp path + std::ofstream(binary) write.
- otf_utils.cc compared ULARGE_INTEGER::QuadPart (ULONGLONG) against a
  signed LONGLONG cast of required_bytes, tripping C4018 (signed/unsigned
  mismatch) as an error. required_bytes is already uint64_t; compare
  unsigned-to-unsigned.

+++ src/otf_utils.cc
@@ -1796,7 +1796,9 @@ std::string ValidateStoragePath(const std::string& path,
 #if defined(_WIN32)
   ULARGE_INTEGER free_bytes;
   if (GetDiskFreeSpaceExW(resolved.c_str(), &free_bytes, nullptr, nullptr)) {
-    if (free_bytes.QuadPart < static_cast<LONGLONG>(required_bytes)) {
+    // QuadPart is ULONGLONG; compare unsigned-to-unsigned (required_bytes is
+    // uint64_t) — a signed LONGLONG cast trips C4018 under /WX.
+    if (free_bytes.QuadPart < required_bytes) {
       return std::string("Not enough free disk space. At least ") +
              (purpose == "cache" ? "100 MB" : "50 MB") + " required.";
     }
